Put the SQL Queries File in the DB Folder and Refactor Accordingly
